Import backup of old webserver from 2015
This is the backup of the lecture demos site on the old physics
department webserver. The webserver was retired in 2015, and while
all other content was migrated to Drupal, the lectore demos went
into a dynamic site named 'DemOnline'. None of the changes that
went into DemOnline are contained in this commit.